SecondView ViewModel 분리
