+++ FreshTomato/Model/Movie.swift
@@ -6,14 +6,20 @@
 //
 
 import Foundation
+import SwiftUI
 
 // MARK: - API 통신으로 데이터 가져와서 저장하는 구조체
+
+
 struct Movie: Identifiable, Codable,Hashable{
     let id: Int
     let title: String
     let backdrop_path: String?
     let overview: String
     let genre_ids: [Int]
+    let release_date: String
+    let vote_average: Double
+   // let runningtime: String
     
     var backdropURL: URL? {
         guard let backdropPath = backdrop_path else { return nil }
