+++ FreshTomato/View/FourthView.swift
@@ -8,12 +8,18 @@
 import SwiftUI
 
 struct FourthView: View {
+    
+    @AppStorage("favoriteMovieTitle") var favoriteMovieTitle: String = ""
+        @AppStorage("favoriteMovieOverview") var favoriteMovieOverview: String = ""
+
+    
     var body: some View {
         NavigationStack {
             List {
                 Section(header: Text("나의 목록")) {
                     
-                    
+                    Text("favariote\(favoriteMovieTitle)")
+                    Text("favariote\(favoriteMovieOverview)")
                 }
                 
                 Section(header: Text("나의 평점")) {
